Refactor TransactionService: add createTransaction, deposit, withdraw, transfer methods and separate Audit, Ledger, Outbox helpers; fix repo typos and add new DTOs

+++ src/main/java/com/projecct/bankx_digital_banking_platform/account/Account.java
@@ -6,6 +6,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.security.SecureRandom;
 
 @Entity
@@ -29,7 +30,7 @@ public class Account {
     @Enumerated(EnumType.STRING)
     private AccountType accountType;
 
-    private Long balance;
+    private BigDecimal balance;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "customer_id", referencedColumnName = "id")
